* commit '7a8909ba812381c60e01f175ec555a2805e53b10':
  CCKEY-282 Removed code smell
  CCKEY-282 Removing code smells
  CCKEY-282 Fixing code smell
  CCKEY-282 Fixing code smells
  CCKEY-282 Adding documentation
  CCKEY-282 Adding job creation on startup
  CCKEY-282 Fixing access on lazy fields
  CCKEY-282 Fixing memory overload

+++ src/main/java/de/communicode/communikey/service/KeyService.java
@@ -209,7 +209,16 @@ public class KeyService {
         String login = SecurityUtils.getCurrentUserLogin();
         User user = userService.validate(login);
         Key key = validate(hashid);
-        UserEncryptedPassword userEncryptedPassword = userEncryptedPasswordRepository.findOneByOwnerAndKey(user, key);
+        UserEncryptedPassword userEncryptedPassword = null;
+        if(user.getAuthorities().contains(authorityService.get(ADMIN))) {
+            userEncryptedPassword = userEncryptedPasswordRepository.findOneByOwnerAndKey(user, key);
+        } else {
+            for (UserGroup userGroup:key.getCategory().getGroups()) {
+                if(user.getGroups().contains(userGroup)) {
+                    userEncryptedPassword = userEncryptedPasswordRepository.findOneByOwnerAndKey(user, key);
+                }
+            }
+        }
         ofNullable(userEncryptedPassword)
             .orElseThrow(UserEncryptedPasswordNotFoundException::new);
         return Optional.of(userEncryptedPassword);
@@ -268,26 +277,29 @@ public class KeyService {
      * @since 0.15.0
      */
     public void removeObsoletePasswords(User user) {
-        userEncryptedPasswordRepository.findAllByOwner(user)
-            .forEach(userEncryptedPassword -> {
-                Key key = userEncryptedPassword.getKey();
-                KeyCategory category = key.getCategory();
-                if (category != null) {
-                    Set<UserGroup> userGroups = category.getGroups();
-                    if (userGroups.isEmpty()) {
-                        deleteUserEncryptedPassword(key, userEncryptedPassword);
-                    } else {
-                        userGroups.forEach(userGroup -> {
-                            if (!user.getGroups().contains(userGroup)) {
-                                deleteUserEncryptedPassword(key, userEncryptedPassword);
-                            }
-                        });
-                    }
+        if (user.getAuthorities().stream()
+            .noneMatch(authority -> authority.getName().equals(ADMIN))) {
+            userEncryptedPasswordRepository.findAllByOwner(user)
+                .forEach(userEncryptedPassword -> {
+                    Key key = userEncryptedPassword.getKey();
+                    KeyCategory category = key.getCategory();
+                    if (category != null) {
+                        Set<UserGroup> userGroups = category.getGroups();
+                        if (userGroups.isEmpty()) {
+                            deleteUserEncryptedPassword(key, userEncryptedPassword);
+                        } else {
+                            userGroups.forEach(userGroup -> {
+                                if (!user.getGroups().contains(userGroup)) {
+                                    deleteUserEncryptedPassword(key, userEncryptedPassword);
+                                }
+                            });
+                        }
 
-                } else if (!key.getCreator().equals(user)){
-                    deleteUserEncryptedPassword(key, userEncryptedPassword);
-                }
-            });
+                    } else if (!key.getCreator().equals(user)) {
+                        deleteUserEncryptedPassword(key, userEncryptedPassword);
+                    }
+                });
+        }
     }
 
     /**
@@ -302,12 +314,15 @@ public class KeyService {
      * @since 0.15.0
      */
     private boolean checkKeyAccess(Key key, KeyPayloadEncryptedPasswords payload) {
-        KeyCategory keyCategory = key.getCategory();
         User user = userService.validate(payload.getLogin());
-        if(user.getAuthorities().contains(authorityService.get(AuthoritiesConstants.ADMIN)))
+        if(user.getAuthorities().contains(authorityService.get(ADMIN))) {
             return true;
+        }
+        KeyCategory keyCategory = key.getCategory();
         for (UserGroup userGroup:keyCategory.getGroups()) {
-            if(user.getGroups().contains(userGroup)) return true;
+            if(user.getGroups().contains(userGroup)) {
+                return true;
+            }
         }
         log.info("User '{}' tried to add an encryptedPassword for user {} without access to the key.", getCurrentUserLogin(), user.getLogin());
         return false;
@@ -342,7 +357,9 @@ public class KeyService {
      */
     private void checkPayloadKeyAccess(Key key, KeyPayload keyPayload) {
         for (KeyPayloadEncryptedPasswords encryptedPasswordsPayload : keyPayload.getEncryptedPasswords()) {
-            if (!checkKeyAccess(key, encryptedPasswordsPayload)) throw new KeyNotAccessibleByUserException();
+            if (!checkKeyAccess(key, encryptedPasswordsPayload)) {
+                throw new KeyNotAccessibleByUserException();
+            }
         }
     }
